Split SFX file upload into its own router to resolve qlty complexity flag
The in-file helper extraction (1254b5b) only brought qlty's file-complexity
count from 56 to 54 — still flagged. Move the upload-specific code (multer
setup, magic-byte detection, filename sanitizing, collision-safe write,
DB-persist-with-rollback) into a new sfxFileUpload.ts router, leaving
sfxFileMutations.ts with just the lighter update/remove routes. Mounted
alongside the other domain routers in sfxMutations.ts. parseWeight, shared
by both files, moves to sfxMutationsShared.ts. Pure refactor — no behaviour
change; the existing helper tests moved with their subject to
sfxFileUpload.test.ts.
